Anexo funcion de auditar y elimino API para crear auditoria elimino carpmeto Middleware
