+++ backend/funciones/funciones.js
@@ -1,4 +1,7 @@
 const { DateTime } = require('luxon');
+const db = require('../database/db.js');
+
+const { validateAudit } = require('../validaciones/validAudit.js');
 
 function obtenerFechayHora(seleccion) {
     const zonaHoraria = 'America/Caracas';
@@ -14,8 +17,22 @@ function obtenerFechayHora(seleccion) {
     }
   }
 
+  function auditar(operacion, id_usuario) {
+
+  try {
+		const fechaActual = obtenerFechayHora("fecha");
+    const horaActual = obtenerFechayHora("hora");
+
+		 db.query("INSERT INTO auditoria (operacion,id_usuario,fecha,hora) VALUES (UPPER($1),$2,CAST($3 as date),CAST($4 as time))",
+			[operacion,id_usuario,fechaActual,horaActual]);
+	} catch (err) {
+		console.error(err.message);
+	}
+	
+  }
 
 
 
 
-  module.exports = {obtenerFechayHora};
+
+  module.exports = {obtenerFechayHora, auditar};
